+++ spec/sandal/enc/a128gcm_spec.rb
@@ -2,8 +2,6 @@ require 'helper'
 require 'openssl'
 require_relative 'shared_examples'
 
-# TODO: These tests are really for the Sandal module rather than just the algorithm -- move them!
-
 if defined? Sandal::Enc::A128GCM
 
 describe Sandal::Enc::A128GCM do