+++ spec/sandal/enc/shared_examples.rb
@@ -6,9 +6,9 @@ shared_examples 'algorithm compatibility' do |enc_class|
   it 'can encrypt and decrypt tokens with the "dir" algorithm' do
     payload = 'Some text to encrypt'
     content_master_key = SecureRandom.random_bytes(32)
-    encrypter = enc_class.new(Sandal::Enc::Alg::Direct.new(content_master_key))
-    token = Sandal.encrypt_token(payload, encrypter)
-    output = Sandal.decode_token(token) { encrypter }
+    enc = enc_class.new(Sandal::Enc::Alg::Direct.new(content_master_key))
+    token = enc.encrypt({}, payload)
+    output = enc.decrypt(token)
     output.should == payload
   end
 
@@ -16,10 +16,9 @@ shared_examples 'algorithm compatibility' do |enc_class|
     payload = 'Some other text to encrypt'
     rsa = OpenSSL::PKey::RSA.new(2048)
     encrypter = enc_class.new(Sandal::Enc::Alg::RSA1_5.new(rsa.public_key))
-    token = Sandal.encrypt_token(payload, encrypter)
-    output = Sandal.decode_token(token) do 
-      enc_class.new(Sandal::Enc::Alg::RSA1_5.new(rsa))
-    end
+    token = encrypter.encrypt({}, payload)
+    decrypter = enc_class.new(Sandal::Enc::Alg::RSA1_5.new(rsa))
+    output = decrypter.decrypt(token)
     output.should == payload
   end
 
@@ -27,10 +26,9 @@ shared_examples 'algorithm compatibility' do |enc_class|
     payload = 'Some more text to encrypt'
     rsa = OpenSSL::PKey::RSA.new(2048)
     encrypter = enc_class.new(Sandal::Enc::Alg::RSA_OAEP.new(rsa.public_key))
-    token = Sandal.encrypt_token(payload, encrypter)
-    output = Sandal.decode_token(token) do 
-      enc_class.new(Sandal::Enc::Alg::RSA_OAEP.new(rsa))
-    end
+    token = encrypter.encrypt({}, payload)
+    decrypter = enc_class.new(Sandal::Enc::Alg::RSA_OAEP.new(rsa))
+    output = decrypter.decrypt(token)
     output.should == payload
   end